+++ NEWS.md
@@ -1,3 +1,5 @@
+# hpa.spatial (development version)
+
 # hpa.spatial 0.2.7
 
 * allow `map_data_with_correspondence()` to take `from_geo` and `to_geo` rather