+++ NEWS.md
@@ -3,6 +3,9 @@
 * allow user to get Modified Monash Model data (for SA1, 2016 ed.) using 
   `get_polygon(name = "mmm19")` or `get_mmm19_poly()`.
 
+* bug fix: pass `value_type` appropriately when `groups` are used in 
+  `map_data_with_correspondence()`.
+
 # hpa.spatial 0.2.7
 
 * allow `map_data_with_correspondence()` to take `from_geo` and `to_geo` rather